Refactor interpreter package out of scalapy.py in preparation for py holding the top-level facade types
