+++ core/native/src/main/scala/me/shadaj/scalapy/interpreter/CPythonAPI.scala
@@ -1,5 +1,4 @@
-package me.shadaj.scalapy.py.interpreter
-
+package me.shadaj.scalapy.interpreter
 
 import scala.scalanative.unsafe._
