+++ core/shared/src/main/scala/me/shadaj/scalapy/py/Any.scala
@@ -3,7 +3,7 @@ package me.shadaj.scalapy.py
 import scala.language.dynamics
 import scala.collection.mutable
 
-import me.shadaj.scalapy.py.interpreter.{CPythonInterpreter, PyValue, VariableReference}
+import me.shadaj.scalapy.interpreter.{CPythonInterpreter, PyValue, VariableReference}
 
 trait Any extends scala.Any { self =>
   private[py] def value: PyValue
