+++ core/shared/src/main/scala/me/shadaj/scalapy/py/Facades.scala
@@ -2,7 +2,7 @@ package me.shadaj.scalapy.py
 
 import scala.language.experimental.macros
 
-import me.shadaj.scalapy.py.interpreter.PyValue
+import me.shadaj.scalapy.interpreter.PyValue
 
 class FacadeValueProvider(private[py] val value: PyValue) extends Any
