+++ core/shared/src/main/scala/me/shadaj/scalapy/py/Global.scala
@@ -2,7 +2,7 @@ package me.shadaj.scalapy.py
 
 import scala.language.dynamics
 
-import me.shadaj.scalapy.py.interpreter.CPythonInterpreter
+import me.shadaj.scalapy.interpreter.CPythonInterpreter
 
 object global extends scala.Dynamic {
   def applyDynamic(method: String)(params: Any*): Dynamic = {
