+++ core/shared/src/main/scala/me/shadaj/scalapy/py/Module.scala
@@ -3,7 +3,7 @@ package me.shadaj.scalapy.py
 import scala.language.dynamics
 import scala.reflect.ClassTag
 
-import me.shadaj.scalapy.py.interpreter.CPythonInterpreter
+import me.shadaj.scalapy.interpreter.CPythonInterpreter
 
 @native trait Module extends Dynamic
