+++ core/shared/src/main/scala/me/shadaj/scalapy/py/package.scala
@@ -4,7 +4,7 @@ import scala.collection.mutable
 
 import scala.concurrent.Future
 
-import me.shadaj.scalapy.py.interpreter.{CPythonInterpreter, PyValue, VariableReference}
+import me.shadaj.scalapy.interpreter.{CPythonInterpreter, PyValue, VariableReference}
 
 package object py {
   def module(name: String) = Module(name)
